Extend ResourceProvider to have resources and aggregates
A ResourceProvider now provides a facade over an Inventory, a
ResourceProviderAggregate and an Allocation. It is possible to
create a ResourcePool, and update its resources and associated
aggregates.

Resources and aggregates are updated under the domain of the object
updating itself (this puts the object in alignment with the state
management done in the resource pools API). It is also possible to
update resources by independently creating Inventory objects with
the correctly associated resource provider. Doing so, however, skips
some of the validation checks that the resource provider does:

* a resource cannot have its `total` or `reserved` changed such that
  the value of `used` would be greater than
  (total-reserved)*allocation_ration.
* a resource cannot be removed if used > 0.

aggregates and resources are loaded lazily

When a ResourceProvider is destroyed the inventories and
ResourceProviderAggregates associated with it are destroyed too
but only if there are no allocations (if used == 0).

A ResourceProviderList now exists with a single classmethod to
get_all(), which will construct ResourceProvider objects for all
ResourceProviders that are not compute nodes. This satisfies the
needs the resource pools api. The disjoint in name may need to
be resolved, if we care.

Partially-Implements: blueprint generic-resource-pools
Change-Id: I5b7dcee81bf37272a012fdfb0c5c8e8c2305045e