Add ResourcePool object
A ResourcePool is a facade over a ResourceProvider, an Inventory, a
ResourceProviderAggregate and an Allocation. It is possible to
create a ResourcePool, assign resources (inventories),
associate aggregates and retrieve resource usage information.

From the standpoint of someone using the API, they'll post up an
id, some aggregates, and some inventories. The errors sent back
to the user-agent can be more granular by:

* creating or getting the resource provider
* getting the aggregates
* validating the provided inventories by creating them
* doing the association of the resource provider and the aggregate
* assembling the response by calculating the resources for
  each inventory

There is a resources property that gets a thing which looks like an
Inventory object but with the used field added.

Change-Id: I5b7dcee81bf37272a012fdfb0c5c8e8c2305045e