Add Allocation and AllocationList objects
Allocation objects are required in the testing and use of
ResourcePool objects, coming in a dependent patch.

An Allocation represents a use of some class of resource from a
particular resource provider by a consumer. They can be created
and destroyed.

It is also possible get a list of allocations associated with a
particular resource provider and resource class.

This also includes a small correction to functional tests of Inventory
so that its structure (where resource_class is defined) is aligned
with the Allocation tests.

Co-Authored-By: Dan Smith <dansmith@redhat.com>
Partially-Implements: blueprint resource-providers-allocations
Change-Id: Iaec8a8e318adfd8feae6496ad962bc02d13adebd