Rudimentary support for gabbi tests in nova
This provides a ConfigFixture that sets up the bare minimum
configuration fixture to load the compute API as a WSGI app,
intercept it using wsgi intercept and run one request to /.

There is no auth used, yet, but we do get a reasonable response back
from /.

Note that a test driven design is being used here to avoid have too
many fixtures and config settings present via the config fixture.
This means that as more requests are added, they will fail because
some service (for example the db) will be missing. This is okay, we
can add the fake services as they are required.

A 'gabbi' target is added to tox.ini. This runs just the tests
generated by nova/tests/functional/gabbi/test_gabbi.py.

gabbi is added to test-requirements.txt

.testr.conf is modified to cause each yaml file in the gabbits
directory to be grouped into its own process when testr runs in
concurrency mode. This is necesary because gabbi use a single
configuration for each test file (the tests in one file are an
ordered suite representing an interaction with the API).

There's a long way to go, but I thought it was worth getting
something to look at and think about.

Change-Id: I5a80237096cc728c185294c5542c01e76e6edea6