Expand tests to use newest gabbi and noauth2 strategy
Newest gabbi has support for sort and filters in jsonpaths.
This makes it possible to query lists of objects in responses,
which is pretty much required in the modern world of unsorted lists
of objects in objects.

The noauth2 strategy lets us fake some keystone handling.

An xfail test has been added for the 204 that the noauth system
returns. 204? Weird.

Change-Id: Iadb3ac9c19b6100d7d73b1b6442e8dafafffe636