WIP resource-pools/placement api
This is a very rough proof of concept of the resource-pools
portion of a new placement api server. It is being done to help
flesh out what is needed for the ResourcePools object.

The api is made up of two apps: One which has no auth and loads up
/ with what will become a basic JSON-Home document. The other
leads to the /resource_pools endpoint. The handler there is a
Selector app dispatching to simple WSGI handler functions.

jsonschema is used to validate the resource pool.

Open questions:

* the jsonschema was not right, so I made one that seems close, but
  maybe not correct
* exception handling needs to be resolved
* at the moment this feels too nova for easy extraction

Next steps:

* continue fulfilling the API and see what that shows

Change-Id: Ided59ecc3726c357ae76ce48251c066a5af20d0e